+++ backend/src/EasyVPN.Application/ConnectionTickets/Queries/GetConnectionTickets/GetConnectionTicketsQuery.cs
@@ -1,4 +1,3 @@
-using EasyVPN.Application.Connections.Queries.GetConfig;
 using EasyVPN.Domain.Entities;
 using ErrorOr;
 using MediatR;