+++ frontend/src/providers/RoutesProvider.tsx
@@ -5,15 +5,18 @@ import { Navigate, Route, Routes, useLocation } from 'react-router-dom';
 
 import { Context } from '..';
 import { Role } from '../api';
+import ClientTicketModal from '../modules/ClientTicketModal';
 import ConfigModal from '../modules/ConfigModal';
 import CreateConnectionModal from '../modules/CreateConnectionModal';
 import DeleteConnectionModal from '../modules/DeleteConnectionModal';
 import ExtendConnectionModal from '../modules/ExtendConnectionModal';
+import PaymentTikcetModal from '../modules/PaymentTikcetModal';
 import AuthPage from '../pages/AuthPage';
 import ClientConnectionsPage from '../pages/ClientConnectionsPage';
 import ForbiddenPage from '../pages/ForbiddenPage';
 import MainPage from '../pages/MainPage';
 import NotFoundPage from '../pages/NotFoundPage';
+import PaymentTicketsPage from '../pages/PaymentTicketsPage';
 import ProfilePage from '../pages/ProfilePage';
 import Root from '../pages/Root';
 
@@ -108,10 +111,12 @@ const RoutesProvider: FC = () => {
                         element={
                             <Auth
                                 with={Role.PaymentReviewer}
-                                for={<>payments tickets</>}
+                                for={<PaymentTicketsPage />}
                             />
                         }
-                    />
+                    >
+                        <Route path=":ticketId" element={<PaymentTikcetModal />} />
+                    </Route>
                 </Route>
                 <Route path="profile" element={<Auth for={<ProfilePage />} />} />
                 <Route
@@ -128,6 +133,7 @@ const RoutesProvider: FC = () => {
                         path=":connectionId/delete"
                         element={<DeleteConnectionModal />}
                     />
+                    <Route path="ticket/:ticketId" element={<ClientTicketModal />} />
                 </Route>
                 <Route path="*" element={<NotFoundPage />} />
             </Route>